Updates from Security Onion Auth repo

+++ so-auth/api/api/tests/test_auth_endpoint.py
@@ -5,9 +5,9 @@ import jwt
 from flask import current_app
 from flask_testing import TestCase
 
-from api import create_app
+from api import create_app, Admin
 from api.models.user import User
-from api.models import db
+from api.models import DB
 from api.tests import SQLALCHEMY_DATABASE_URI
 
 
@@ -24,25 +24,35 @@ class TestAuth(TestCase):
         return app
 
     def setUp(self) -> None:
-        db.create_all()
+        DB.create_all()
+
+        # check if admin entry already exists and if not, add it
+        if not Admin.query.filter_by(created=True).first():
+            admin_instance = Admin()
+            DB.session.add(admin_instance)
+            DB.session.commit()
 
         user = User('test_username', 'test_password')
-        db.session.add(user)
-        db.session.commit()
-        assert user in db.session
+        DB.session.add(user)
+        DB.session.commit()
+        assert user in DB.session
 
         self.valid_jwt_auth = User.encode_token(user.id, user.username)
-        self.valid_jwt_refresh = User.encode_token(user.id, user.username, is_refresh=True)
+        self.valid_jwt_refresh = User.encode_token(
+            user.id, user.username, is_refresh=True)
         self.secret_key = current_app.config.get("SECRET_KEY")
 
     def tearDown(self) -> None:
-        db.session.remove()
-        db.drop_all()
-        os.remove('db.testing.sqlite')
+        DB.session.remove()
+        DB.drop_all()
+        if os.path.exists('db.testing.sqlite'):
+            os.remove('db.testing.sqlite')
 
     @classmethod
     def tearDownClass(cls):
         try:
+            if os.path.exists('db.testing.sqlite'):
+                os.remove('db.testing.sqlite')
             os.remove('secret')
         except FileNotFoundError:  # pragma: no cover
             pass
@@ -64,15 +74,6 @@ class TestAuth(TestCase):
         ))
 
         self.assert200(response)
-
-    def test_register_remote(self):
-        response = self.client.post(
-            '/auth/register',
-            json=dict(username='test_username_new', password='super_secret_password'),
-            environ_base={'REMOTE_ADDR': '192.168.1.1'}
-        )
-
-        self.assert401(response)
 
     def test_register_bad_password(self):
         response = self.client.post('/auth/register', json=dict(
@@ -145,7 +146,8 @@ class TestAuth(TestCase):
             password='test_password'
         ))
 
-        self.client.set_cookie('localhost', 'Auth-Token', self.valid_jwt_auth + b'a')
+        self.client.set_cookie('localhost', 'Auth-Token',
+                               self.valid_jwt_auth + b'a')
         response = self.client.post('/auth/')
 
         self.assert401(response)
@@ -178,7 +180,7 @@ class TestAuth(TestCase):
 
         response = self.client.post('/auth/')
 
-        self.assert500(response)
+        self.assert401(response)
 
     def test_renew_token(self):
         self.client.post('/auth/login', json=dict(
@@ -187,7 +189,8 @@ class TestAuth(TestCase):
             remember_me='true'
         ))
 
-        response = self.client.post('/auth/renew', headers={'X-Original-URI': 'test_uri'})
+        response = self.client.post(
+            '/auth/renew', headers={'X-Original-URI': 'test_uri'})
 
         self.assertRedirects(response, 'auth/test_uri')
 
@@ -201,4 +204,3 @@ class TestAuth(TestCase):
         response = self.client.post('/auth/renew')
 
         self.assert200(response)
-
